PODS-2432: controller moved to correct location
